Separated to interface and implementation
