+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/services/AntibioticsService.java
@@ -2,58 +2,15 @@ package fi.tuni.koodimankelit.antibiootit.services;
 
 import java.util.List;
 
-import org.springframework.stereotype.Service;
-
 import fi.tuni.koodimankelit.antibiootit.database.data.DiagnoseInfo;
-import fi.tuni.koodimankelit.antibiootit.models.AntibioticTreatment;
-import fi.tuni.koodimankelit.antibiootit.models.DosageFormula;
-import fi.tuni.koodimankelit.antibiootit.models.DosageResult;
-import fi.tuni.koodimankelit.antibiootit.models.Instructions;
-import fi.tuni.koodimankelit.antibiootit.models.Measurement;
 import fi.tuni.koodimankelit.antibiootit.models.Parameters;
-import fi.tuni.koodimankelit.antibiootit.models.Treatment;
 import fi.tuni.koodimankelit.antibiootit.models.Treatments;
 
-@Service
-public class AntibioticsService {
 
-    private final DataHandler dataHandler;
+public interface AntibioticsService {
 
-    public AntibioticsService(DataHandler dataHandler) {
-        this.dataHandler = dataHandler;
-    }
 
-    public Treatments calculateTreatments(Parameters parameters) {
-        return this.mockCalculator();
-    }
+    public Treatments calculateTreatments(Parameters parameters);
 
-    public List<DiagnoseInfo> getAllDiagnoseInfos() {
-        return this.dataHandler.getAllDiagnoseInfos();
-    }
-    
-    // Mock calculation of treatments. Might be useful for unit testing also
-    private Treatments mockCalculator() {
-
-        Treatments d =  new Treatments("J03.0", "Streptokokki A");
-        Treatment t = new Treatment();
-        d.addTreatment(t);
-
-        t.addAntibiotic(
-            new AntibioticTreatment(
-                "Mikstuura",
-                "infoteksti...",
-                "Amoksisilliini",
-                new Instructions(10, 3),
-                new DosageFormula(
-                    new Measurement("mg/ml", 100),
-                    new Measurement("mg/kg/vrk", 50)),
-                new DosageResult(
-                    new Measurement("ml", 4)
-                )
-            )
-        );
-
-        return d;
-    }
-
+    public List<DiagnoseInfo> getAllDiagnoseInfos();
 }
